Move Metal profiling scope out of main
